+++ Editor/Folders.cs
@@ -1,0 +1,67 @@
+using UnityEngine;
+using System.IO;
+using static System.IO.Path;
+using static UnityEditor.AssetDatabase;
+
+namespace RAStudio.SetupTools.Editor
+{
+    /// <summary>
+    /// Provides methods for creating moving and deleting folders in the Unity project.
+    /// </summary>
+    public static class Folders
+    {
+        public static void Create(string root, params string[] folders)
+        {
+            var fullpath = Combine(Application.dataPath, root);
+            if (!Directory.Exists(fullpath))
+            {
+                Directory.CreateDirectory(fullpath);
+            }
+
+            foreach (var folder in folders)
+            {
+                CreateSubFolders(fullpath, folder);
+            }
+        }
+
+        static void CreateSubFolders(string rootPath, string folderHierarchy)
+        {
+            var folders = folderHierarchy.Split('/');
+            var currentPath = rootPath;
+
+            foreach (var folder in folders)
+            {
+                currentPath = Combine(currentPath, folder);
+                if (!Directory.Exists(currentPath))
+                {
+                    Directory.CreateDirectory(currentPath);
+                }
+            }
+        }
+
+        public static void Move(string newParent, string folderName)
+        {
+            var sourcePath = $"Assets/{folderName}";
+            if (IsValidFolder(sourcePath))
+            {
+                var destinationPath = $"Assets/{newParent}/{folderName}";
+                var error = MoveAsset(sourcePath, destinationPath);
+
+                if (!string.IsNullOrEmpty(error))
+                {
+                    Debug.LogError($"Failed to move {folderName}: {error}");
+                }
+            }
+        }
+
+        public static void Delete(string folderName)
+        {
+            var pathToDelete = $"Assets/{folderName}";
+
+            if (IsValidFolder(pathToDelete))
+            {
+                DeleteAsset(pathToDelete);
+            }
+        }
+    }
+}